don't automatically call session_start() in render_action() because it makes php send headers to disable caching, and in general we're not going to need sessions everywhere and php is probably using server-side temporary files for sessions.
introduce a static $session variable that can be set per-controller (defaulting to "off") and takes arguments like rails' old session function (like $session = array("on", "except" => array("foo"))) which will determine whether sessions are enabled.

any time a form helper is used with a post method and the xsrf protection is enabled, explicitly call session_start() since we'll need it to store and retrieve the token.

it would be helpful if we could lazy-load sessions like rails does now (with a strictly cookie-based store like rails, too), but since php only calls session_start() when using session_register() and not when using $_SESSION, and session_register() is now deprecated (wtf), we have to require the user to turn them on or off.