automatic etag support
md5() the output buffer before sending it back to the client and if it matches
a If-None-Match header sent, send no data and just give a 304 header.  this may
need a knob to be disabled on long, dynamic output to avoid useless md5() of
the data.

also set caching headers properly depending on whether sessions are being used.