actually make use of php-ar environments; now multiple-db setups work
instead of just using the development environment for php-ar, use
what is configured in the db.ini file but set the default to
HALFMOON_ENV's value, so that we will default to development or
whatever we're in, like we did before.

now that php-ar has the full list of db configs, models can use:

    static $connection = "blah";

and talk to the connection specified by that config in db.ini rather
than the default.  this allows different models to talk to different
databases.  verified by loading data from mysql and postgresql
databases at the same time.