introduce global function raw() to mark text as not-to-be-escaped
to remove ambiguity over what parts of built forms are escaped and
and which are not, and increase security in everything else that
prints html, run everything supplied by the application through h()
to escape html.

since there may be instances where html should be printed, like in
flash_errors(), introduce a global function raw() that can be used
to protect a string until it hits the helpers, where it will be
recognized as a raw object and not run through h().  most things
that did h() now do raw_or_h().

note that the raw object must continue all the way through to the
helper, so doing

    $this->add_flash_error(raw("<b>error:</b> ") . $somevar)

does nothing, because the Raw object will get __toString()'d right
away and get seen as a normal string when it gets to the helper,
which will run h() on it.  the entire argument must be the raw
object with your own html protection done on the parts you need it:

    $this->add_flash_error(raw("<b>error:</b> " . h($somevar)))